Let there be config
Implement configuration based on TOML and wire it up to the various
config structs. Start writing the first command to interact with the
system: adding a population. Some smaller changes include changing the
type of Selector/Evaluator.MachineRun to bool and removing the last bits
of configurable cell boundaries that were lurking