Checkpoint check in
Rework the generation engine slightly to not have a single output
channel and one aggregate batch for persisting. Instead, feed a
UnitPersistor func to the processor to push when a stack of units are
ready to be persisted so that persisting happens within multiple
goroutines.

Also rework Population a little bit to contain the logic for processing
a generating.

Disable the noisy gorm logs.

Finally, tweak SQLite pragmas to actually use threads, and enable
mmapping. This helps a lot with performance.